feat(crypto): Add method to make rng secure.
Libsodium has to initiallized to be threadsafe. Moved all crypto into it's own module and reexport
important bits of libsodium.

KeyPair is now part of crypto mod.
